refactor(eighth): move Celery tasks into a separate module
Also begin fixing #731
