+++ Ion.egg-info/SOURCES.txt
@@ -208,6 +208,7 @@ intranet/apps/eighth/exceptions.py
 intranet/apps/eighth/models.py
 intranet/apps/eighth/notifications.py
 intranet/apps/eighth/serializers.py
+intranet/apps/eighth/tasks.py
 intranet/apps/eighth/tests.py
 intranet/apps/eighth/urls.py
 intranet/apps/eighth/utils.py
