+++ intranet/apps/eighth/views/admin/scheduling.py
@@ -17,7 +17,7 @@ from ...forms.admin.blocks import BlockSelectionForm
 from ...forms.admin.scheduling import ScheduledActivityForm
 from ...models import EighthActivity, EighthBlock, EighthRoom, EighthScheduledActivity, EighthSponsor, EighthSignup
 from ...utils import get_start_date
-from ...notifications import room_changed_single_email
+from ...tasks import room_changed_single_email
 from ....auth.decorators import eighth_admin_required
 from .....utils.serialization import safe_json
